LayoutTests/fast/frames/seamless/seamless-inline-expected.txt

FAIL window.getComputedStyle(iframe1).width should be 150px. Was 0px.
FAIL window.getComputedStyle(iframe1).height should be 50px. Was 104px.
FAIL window.getComputedStyle(parent1).height should be 50px. Was 108px.
FAIL window.getComputedStyle(iframe2).width should be 100px. Was 0px.
FAIL window.getComputedStyle(iframe2).height should be 100px. Was 104px.
FAIL window.getComputedStyle(parent2).height should be 100px. Was 108px.
